create map folder mode

+++ src/electron/makeNewMap.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const { makeMapFile } = require('./loadFuncs')
 
-exports.makeNewMap = async (dir) => {
+exports.makeNewMap = async (dir, folderMode) => {
   const mapId = dir.split(/[\\/]/gm).pop()
 
   await makeRootFiles(dir, mapId)
@@ -11,11 +11,12 @@ exports.makeNewMap = async (dir) => {
   await makeLocaleDir(dir, mapId)
   await makeParticlesDir(dir, mapId)
   await makeScriptsDir(dir, mapId)
-  await makeMapFile(dir, dir + ".map")
-
-  await fs.promises.rm(dir, { recursive: true, force: true })
+  
+  if(!folderMode){
+    await makeMapFile(dir, dir + ".map")
+    await fs.promises.rm(dir, { recursive: true, force: true })
+  }
 }
-
 
 const makeRootFiles = async (dir, mapId) => {
   await fs.promises.writeFile(
